Add obsolete intermediary version of Hydra ("HydraV").
This version was put together by Min Deng and was a reworking of
Dr. McUmber's original Perl version to use a Visitor design pattern.  Karli
Lopez used this version as a guide when he built the new Java version that
also uses a Visitor pattern (hence the eventual name "HydraJV").