Revise the way signals/operations are handled to generate safer Promela code.
Hydra could have been introducing feature interactions by creating an
mtype for each signal.  If two classes define a signal (operation)
called "start", then transitions in both classes could be triggered
when that signal is invoked on just one of them.  Sometimes shared
triggers are the intended behavior, but this issue could create
unexpected and hard-to-debug behavior.  To solve the problem, I
mangled the signal names to include the containing class name.