Improve tracking of attributes that are instances of classes.
The old implementation was causing problems due to timing.  That is,
an instance of class X might be declared before X itself had been
declared and parsed.  The new implementation marks each attribute that
is an instance of a class and then processes them when the HIL parser
has finished.